[FRONT&BACK] Secured endpoint for a better security

+++ api/src/main/java/com/finalproject/sheeks/services/AnnounceService.java
@@ -58,7 +58,11 @@ public class AnnounceService implements IAnnounceService{
     @Override
     public void deleteAnnounce(Long id) {
         Announce announceToDelete = announceRepository.findAnnounceById(id);
-        announceRepository.delete(announceToDelete);
+
+        if(announceToDelete.getUser() == userService.getLoggedUser()) {
+            announceRepository.delete(announceToDelete);
+        }
+
     }
